feat: restructure team and member routes & controllers
