+++ backend/package.json
@@ -21,10 +21,7 @@
   },
   "dependencies": {
     "@prisma/client": "^5.11.0",
-    "axios": "^1.6.8",
     "cors": "^2.8.5",
-    "express": "^4.19.1",
-    "jsonwebtoken": "^9.0.2",
-    "siwe": "^2.1.4"
+    "express": "^4.19.1"
   }
 }
