+++ backend/src/index.ts
@@ -1,7 +1,8 @@
 import cors from "cors";
 import express from "express";
 import { PrismaClient } from "@prisma/client";
-import router from "./routes/index";
+import teamRoutes from "./routes/teamRoutes";
+import memberRoutes from "./routes/memberRoutes";
 
 const prisma = new PrismaClient();
 
@@ -9,8 +10,8 @@ const app = express();
 app.use(express.json());
 app.use(cors());
 
-app.use(router);
-
+app.use(teamRoutes);
+app.use(memberRoutes);
 const port = parseInt(process.env.PORT as string) || 3000;
 app.listen(port, () => {
   console.log(`helloworld: listening on port ${port}`);
